stress/requirements.txt
PyYAML==6.0
prometheus-client==0.17.0
